requirements.txt
flask==2.3.3
diffusers>=0.27.0
huggingface_hub>=0.23.0,<1.0
torch==2.2.2
transformers>=4.37.0
safetensors==0.4.2
pillow==10.3.0
psutil==5.9.8
numpy==1.26.4
accelerate==0.29.3
